docker/training/requirements.txt
python-dotenv
joblib
tensorflow
kaggle